+++ changelog.md
@@ -1,5 +1,10 @@
 # MODX Vapor Changelog
 
+- Add excludeExtraTablePrefix to avoid prepending table_prefix for specific tables
+- Add excludeExtraTables option for excluding specific extra non-core tables
+- Add excludeFiles option for excluding specific files/dirs from base_path
+- Add ability to provide vaporOptions from config file
+
 ## 1.0.0-beta-3 (September 26, 2012)
 
 - Fix PHP warning when table_prefix is empty